refetch envirements after promote step happened
