+++ web/client/src/workers/sqlglot/worker.ts
@@ -1,6 +1,6 @@
 declare function importScripts(...urls: string[]): void
 
-importScripts('https://cdn.jsdelivr.net/pyodide/v0.22.1/full/pyodide.js')
+importScripts('https://cdn.jsdelivr.net/pyodide/v0.23.2/full/pyodide.js')
 
 const scope = self as any
