+++ web/server/console.py
@@ -99,8 +99,7 @@ class ApiConsole(TerminalConsole):
         self.queue.put_nowait(self._make_event(data, event="tests", ok=ok))
 
     def log_success(self, msg: str) -> None:
-        self.queue.put_nowait(self._make_event(msg))
-        self.stop_snapshot_progress()
+        self.queue.put_nowait(self._make_event(msg, event="promote-environment"))
 
     def log_exception(self) -> None:
         """Log an exception."""
